refactor(master): clean up code
