+++ master/api_service.go
@@ -2100,7 +2100,7 @@ func (m *Server) updateVol(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if req.storeMode, req.mpLayout, err = parseRocksDbFieldToUpdateVol(r, vol); err != nil {
+	if req.storeMode, err = parseRocksDbFieldToUpdateVol(r, vol); err != nil {
 		sendErrReply(w, r, &proto.HTTPReply{Code: proto.ErrCodeParamError, Msg: err.Error()})
 		return
 	}
@@ -2122,7 +2122,6 @@ func (m *Server) updateVol(w http.ResponseWriter, r *http.Request) {
 	newArgs.txConflictRetryInterval = req.txConflictRetryInterval
 	newArgs.txOpLimit = req.txOpLimit
 	newArgs.enableQuota = req.enableQuota
-	newArgs.MpLayout = req.mpLayout
 	newArgs.DefaultStoreMode = proto.StoreMode(req.storeMode)
 	if req.coldArgs != nil {
 		newArgs.coldArgs = req.coldArgs
@@ -2186,34 +2185,6 @@ func (m *Server) volExpand(w http.ResponseWriter, r *http.Request) {
 	}
 
 	msg = fmt.Sprintf("update vol[%v] successfully\n", name)
-	sendOkReply(w, r, newSuccessHTTPReply(msg))
-}
-
-func (m *Server) setVolConvertTaskState(w http.ResponseWriter, r *http.Request) {
-	var (
-		err      error
-		newState int
-		name     string
-		authKey  string
-		msg      string
-	)
-
-	if name, authKey, newState, err = parseRequestToSetVolConvertSt(r); err != nil {
-		sendErrReply(w, r, &proto.HTTPReply{Code: proto.ErrCodeParamError, Msg: err.Error()})
-		return
-	}
-
-	if newState < int(proto.VolConvertStInit) || newState > int(proto.VolConvertStFinished) {
-		err = fmt.Errorf("unknown state:%d", newState)
-		sendErrReply(w, r, newErrHTTPReply(err))
-		return
-	}
-
-	if err = m.cluster.setVolConvertTaskState(name, authKey, proto.VolConvertState(newState)); err != nil {
-		sendErrReply(w, r, newErrHTTPReply(err))
-		return
-	}
-	msg = fmt.Sprintf("Vol[%v] convert task state change to be [%v] successfully\n", name, newState)
 	sendOkReply(w, r, newSuccessHTTPReply(msg))
 }
 
@@ -2350,10 +2321,6 @@ func (m *Server) checkCreateReq(req *createVolReq) (err error) {
 
 	if !(req.storeMode == (proto.StoreModeMem) || req.storeMode == (proto.StoreModeRocksDb)) {
 		return fmt.Errorf("storeMode can only be %d and %d,received storeMode is[%v]", proto.StoreModeMem, proto.StoreModeRocksDb, req.storeMode)
-	}
-
-	if req.layout.PercentOfMP > 100 || req.layout.PercentOfReplica > 100 || req.layout.PercentOfMP < 0 || req.layout.PercentOfReplica < 0 {
-		return fmt.Errorf("mpPercent repPercent can only be [0-100],received is[%v - %v]", req.layout.PercentOfMP, req.layout.PercentOfReplica)
 	}
 	return nil
 }
@@ -2540,8 +2507,6 @@ func newSimpleView(vol *Vol) (view *proto.SimpleVolView) {
 		Forbidden:               vol.Forbidden,
 		EnableAuditLog:          vol.EnableAuditLog,
 		DefaultStoreMode:        vol.DefaultStoreMode,
-		MpLayout:                vol.MpLayout,
-		ConvertState:            vol.convertState,
 	}
 
 	vol.uidSpaceManager.RLock()
